Predict wielding
Predicted do_afters means stonks are up.
