--- Content.Server/Wieldable/Components/IncreaseDamageOnWieldComponent.cs
@@ -1,11 +1,0 @@
-﻿using Content.Shared.Damage;
-
-namespace Content.Server.Wieldable.Components
-{
-    [RegisterComponent, Access(typeof(WieldableSystem))]
-    public sealed class IncreaseDamageOnWieldComponent : Component
-    {
-        [DataField("damage", required: true)]
-        public DamageSpecifier BonusDamage = default!;
-    }
-}
