+++ Content.Shared/Wieldable/BeforeUnwieldEvent.cs
@@ -1,0 +1,5 @@
+namespace Content.Shared.Wieldable;
+
+public sealed class BeforeUnwieldEvent : CancellableEntityEventArgs
+{
+}
