+++ Content.Shared/Wieldable/Components/IncreaseDamageOnWieldComponent.cs
@@ -1,0 +1,10 @@
+using Content.Shared.Damage;
+
+namespace Content.Shared.Wieldable.Components;
+
+[RegisterComponent, Access(typeof(WieldableSystem))]
+public sealed class IncreaseDamageOnWieldComponent : Component
+{
+    [DataField("damage", required: true)]
+    public DamageSpecifier BonusDamage = default!;
+}
